+++ test/actions/down.test.js
@@ -151,28 +151,6 @@ describe("down", () => {
     expect(migration.down.called).to.equal(true);
   });
 
-  it("should be able to downgrade callback based migration that has both the `db` and `client` arguments", async () => {
-    migration = {
-      down(theDb, theClient, callback) {
-        return callback();
-      }
-    };
-    migrationsDir = mockMigrationsDir();
-    down = loadDownWithInjectedMocks();
-    await down(db, client);
-  });
-
-  it("should be able to downgrade callback based migration that has only the `db` argument", async () => {
-    migration = {
-      down(theDb, callback) {
-        return callback();
-      }
-    };
-    migrationsDir = mockMigrationsDir();
-    down = loadDownWithInjectedMocks();
-    await down(db);
-  });
-
   /* eslint no-unused-vars: "off" */
   it("should allow downgrade to return promise", async () => {
     migrationsDir = mockMigrationsDir();